eleventy: add permalinks to all pages
We add `permalink` to all the pages and align their syntax for
consistency.

The `permalink` tells eleventy how to output these files into the build
directory.  For example, if you have a file in the root of `src` called
`about.njk` without a `permalink` it will be built as `/about/index.html`.

This is not intuitive, IMO, so I update each of the templates to have
the permalink which lets me control how they are out.

I can imagine the default is actually desireable, but I'm not sure I
understand why ATM.